add support for object.__new__ (base patch provided by Chris Torek)
The builtin object.__new__ function takes a class and returns
an object of that type, i.e., we should infer the result has
whatever type we can find from the first argument.

To make this work (with Python2.5 at least), we need to note
that built in functions whose __module__ is None are really
in the __builtin__ module.